Update to sentence feedback.
Sentence feedback still occurs for entire word/sentence based on current character's correctness but the validity is now local to the current word rather than impacting the rest of the sentence. This fixes an issue where if you typed one too many or few letters in a word, it would impact the feedback of the rest of the sentence. The validity is now based on the locality of each respective word that is typed when compared to the current word to be typed.

Last commit also fixed a bug where wpm was calculated incorrectly.