Fix to insertAttempt.
NOTE: this now uses username to get email to get ID to update the attempts data. This means that it is expecting usernames to always be unique. So we can change this or make sure usernames are unique. I did this because I was struggling to get email from authStore.user or something similar in authStore as authStore.user is now just username.